+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## \[Unreleased\]
+- Add tabular data import/export
+  (<https://github.com/openvinotoolkit/datumaro/pull/1089>)
 
 ## 11/07/2023 - Release 1.4.0rc2
 ### New features